requirements.txt
numpy==1.17.0
tensorflow-gpu==1.12.0
typing-extensions==3.7.4.1
typing-inspect==0.5.0
tqdm==4.31.1
pickleshare==0.7.5
scikit-learn==0.20.3
scipy==1.2.1
pandas==0.24.2
matplotlib==2.0.2
seaborn==0.9.0
dataclasses==0.7
dataclasses-json==0.3.5
Click==7.0
fs==2.4.4
six==1.11.0
git+https://github.com/thu-ml/zhusuan.git@48c0f4e
git+https://github.com/haowen-xu/tfsnippet.git@v0.2.0-alpha4
git+https://github.com/haowen-xu/ml-essentials.git
